% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerMethod.R
\name{power.method}
\alias{power.method}
\title{Implementing the Power Method}
\usage{
power.method(A, v = NULL, epsilon = 1e-06, max_iter = 100, plot = FALSE)
}
\arguments{
\item{A}{square numeric matrix}

\item{v}{numeric vector with length equal to the number of rows of A}

\item{epsilon}{desired maximum of approximation}

\item{max_iter}{maximum number of iterations before outputting}

\item{plot}{TRUE/FALSE plot or not}
}
\value{
power.method() returns a list which includes a vector of errors at each iteration, a vector of the approximate eigenvector at each iteration and the dominant eigenvector of A
}
\description{
Given a square matrix A, implement the Power Method to approximate the dominant eigenvector of A up to a specificed error
}
\examples{
A = rbind(c(1,3),4,5)
power.method(A)
eigen(A)

}
