% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerMethod.R
\name{power.method}
\alias{power.method}
\title{Implementing the Power Method}
\usage{
power.method(
  A,
  dominant = TRUE,
  v = NULL,
  epsilon = 1e-06,
  max_iter = 100,
  plot = FALSE
)
}
\arguments{
\item{A}{square numeric matrix}

\item{dominant}{TRUE/FALSE choose to find the dominant or minimal eigenvector}

\item{v}{numeric vector with length equal to the number of rows of A}

\item{epsilon}{desired maximum of approximation}

\item{max_iter}{maximum number of iterations before outputting}

\item{plot}{TRUE/FALSE choose to plot the error versus iterations}
}
\value{
power.method() returns a list which includes a vector of errors at each iteration, vectors of the approximate eigenvectors/eigenvalues at each iteration, and the dominant/minimal eigenvector and eigenvalue of A
}
\description{
Given a square matrix A, implement the Power Method to approximate the dominant or minimal eigenvector of A up to a specified error
}
\examples{
A = rbind(c(1,3),4,5)
power.method(A)
eigen(A)

}
